src/test/java/com/bundle/case1/TripletFilterTest.java
//package com.bundle.case1;
//
//import com.bundle.case1.utility.TripletFilter;
//import org.junit.jupiter.api.BeforeEach;
//import org.junit.jupiter.api.Test;
//
//import java.util.Arrays;
//import java.util.Collections;
//import java.util.List;
//
//import static org.junit.jupiter.api.Assertions.assertEquals;
//
//public class TripletFilterTest {
//    private TripletFilter tripletFilter;
//
//    @BeforeEach
//    public void setUp() {
//        tripletFilter = new TripletFilter();
//    }
//
//    @Test
//    public void testFilterTriplets_withNullInput() {
//        List<Integer> expectedOutput = Collections.emptyList();
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(null);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_withEmptyInput() {
//        List<Integer> inputData = Collections.emptyList();
//        List<Integer> expectedOutput = Collections.emptyList();
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_withInsufficientNumbers() {
//        List<Integer> inputData = Arrays.asList(10, 20);
//        List<Integer> expectedOutput = Collections.emptyList();
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_tripletsBelowThreshold() {
//        List<Integer> inputData = Arrays.asList(10, 20, 30, 15, 15, 15);
//        List<Integer> expectedOutput = Collections.emptyList();
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_tripletsAtThreshold() {
//        List<Integer> inputData = Arrays.asList(30, 30, 30);
//        List<Integer> expectedOutput = Arrays.asList(30, 30, 30);
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_tripletsAboveThreshold() {
//        List<Integer> inputData = Arrays.asList(50, 50, 50);
//        List<Integer> expectedOutput = Arrays.asList(50, 50, 50);
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_multipleTripletsMixedSums() {
//        List<Integer> inputData = Arrays.asList(
//                30, 30, 30,
//                10, 20, 30,
//                40, 50, 60
//        );
//        List<Integer> expectedOutput = Arrays.asList(30, 30, 30, 40, 50, 60);
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//
//    @Test
//    public void testFilterTriplets_withNegativeNumbers() {
//        List<Integer> inputData = Arrays.asList(-50, -25, -25, 50, 25, 25);
//        List<Integer> expectedOutput = Arrays.asList(50, 25, 25);
//
//        List<Integer> actualOutput = tripletFilter.filterTriplets(inputData);
//
//        assertEquals(expectedOutput, actualOutput);
//    }
//}
